feat(sdk): Wait for transaction receipts
Using a local environment with testrpc, transactions happen so quickly, you often don't need to wait
for the receipt before the results are written onto the chain. However, when interacting with a
remote chain, it's necessary to delay sequential transactions until the receipt for previous
transactions are available. I also updated some docblocks :)