refactor(explorer): Added a TransactionStatus class and restructured the TransactionsContainer
TransactionStatus no longer includes the full transaction object/receipt since this information can
be queried for on-chain with the transaction hash. There is now an enhancer that allows a mutation
to be bound to a TransactionStatus. This is really convenient for keeping the actual rpc method
decoupled from application state, while still allowing the convenience of not manually wiring the
two together each time you want a mutation to automatically know about and update a
TransactionStatus.