Updated the default frame-size for iframe
The previous version provided a user with an iframe of 240x160 and also set the aspect ratio at 16:9. This is inconsistent as 240x160 represents an aspect ratio of 4:3.

This change to providing a user with an iframe of 640x360 is consistent with an aspect ratio of 16:9 which is the "standard" aspect ratio for e.g. 1080p, 720p, 4k (3840x2160) etc.