Fix production build
The production build failing was the result of a bug related to an older version react-scripts not playing nice with the latest version of graphql (https://github.com/graphql/graphql-js/issues/1248).

Updating to the latest version of react-scripts (v2.1.1) fixed the bug. It also reduces the production build *considerably* (from over 1mb down to 48kb).