Remove unnecessary npx calls and dependencies
This commit adds a "depcheck" script that checks we don't have any
unused or missing dependencies. There are still a few false negatives
so I haven't added it to the linting process yet.

It also removes a lot of unnecessary usage of `npx`, in the `scripts`
field of `package.json` — `npm run` or `yarn run` already does the same
thing `npx` does, executing the local versions of packages.